working. waiting for instance to be interrupted

+++ spot-manager.py
@@ -28,6 +28,10 @@ logger = logging.getLogger(__name__)
 # ec2 client for spot requests
 client = boto3.client('ec2')
 
+# route 3 client
+route53_client = boto3.client('route53')
+
+
 # sqs client
 sqs = boto3.client('sqs',
                    aws_access_key_id     = os.environ['AWS_ACCESS_KEY_ID'],
@@ -46,8 +50,8 @@ def log(msg):
     # print ('{} - {}'.format(time.strftime("%x %H:%M:%S"), msg))
 
 
-def request_new_spot():
-    print('TODO: request new spot')
+# def request_new_spot():
+#     print('TODO: request new spot')
 
 
 def update_instance_state(iid):
@@ -72,7 +76,7 @@ def delete_message(receipt_handle):
 # update_instance_record(body['name'], body['price'], body['ami'], body['type'], body['sg'], body['subnet'])
 #
 
-def update_instance_record(name, sir, price, ami, _type, sg, subnet, ns_record=""):
+def update_instance_record(name, sir, price, ami, _type, sg, subnet, ns_record):
     query = Instance.select().where(Instance.name == name)
     if query.exists():
         logger.info('updating record name: {}, with sir: {} params.'.format(name, sir))
@@ -83,6 +87,7 @@ def update_instance_record(name, sir, price, ami, _type, sg, subnet, ns_record="
         instance.type = _type
         instance.sg = sg
         instance.subnet = subnet
+        instance.ns_record = ns_record
         instance.save()
     else:
         logger.info('adding new record: name: {} with sir: {} params.'.format(name, sir))
@@ -92,12 +97,12 @@ def update_instance_record(name, sir, price, ami, _type, sg, subnet, ns_record="
 
 
 def add_message_record(source, text, sent, body):
-    logger.info('adding new record to Message table [from:{}, text{}]'.format(source, text))
+    logger.info('adding new record to Message table [from:{}, text:{}]'.format(source, text))
     message = Message.create(source=source, text=text, sent=sent, body=body)
     message.save()
 
 
-#- {'text': 'New spot requested from spot-manager', 'price': '0.05', 'ami': 'ami-0fc8c8e37cd7db658', 'type': 'm5a.large', 'sg': 'sg-056964e89bbf05266', 'subnet': 'subnet-03fb65f37827a8971'}
+#- {'text': 'New spot requested from web-server', 'price': '0.05', 'ami': 'ami-0fc8c8e37cd7db658', 'type': 'm5a.large', 'sg': 'sg-056964e89bbf05266', 'subnet': 'subnet-03fb65f37827a8971'}
 
 def handle_incomming_message(body, receipt_handle, timestamp):
 
@@ -115,18 +120,31 @@ def handle_incomming_message(body, receipt_handle, timestamp):
     if 'text' in body.keys():
         log('got a Lumos message from topic.')
         if body['text'] == 'Lumos new spot instance launched':
-            logger.info('new spot instance launched: {}.'.format(body['iid']))
             add_message_record(source="lumos", text=body['text'], sent=timestamp, body=body)
-            logger.info('deleting message.')
-            delete_message(receipt_handle)
+            logger.info('new spot instance launched: {}.'.format(body['iid']))
+            query = Instance.select().where(Instance.iid == body['iid'])
+            if query.exists():
+                instance = Instance.select().where(Instance.iid == body['iid'])[0]
+                instance.public_ip = body['public_ip']
+                logger.info('updating record for: {} with public_ip: {}.'
+                            .format(instance.name, instance.public_ip))
+                instance.save()
+                logger.info('updating route53 for: {} [ns_record: {} public_ip: {}].'
+                            .format(instance.name, instance.ns_record, instance.public_ip))
+                resp = change_ns_record(instance.name, instance.ns_record, instance.public_ip)
+                logger.info('deleting message.')
+                delete_message(receipt_handle)
+            else:
+                logger.info('ignoring new spot with iid: {} (not managed by spot-manager).'.format(body['iid']))
 
-        elif body['text'] == 'New spot requested from spot-manager':
+        elif body['text'] == 'New spot requested from web-server':
             logger.info('new spot request for {} : (price:{}, ami: {}, type: {}, sg: {}, subnet:{})'
                         .format((body['name']), (body['price']), format(body['ami']), format(body['type']),format(body['sg']),format(body['subnet'])))
-            resp = request_new_spot(body['price'], body['ami'], body['type'], body['sg'], body['subnet'])
+            resp = request_new_spot(body['name'], body['price'], body['ami'], body['type'], body['sg'], body['subnet'])
             # get the spot req id
             sir = resp['SpotInstanceRequests'][0]['SpotInstanceRequestId']
-            update_instance_record(body['name'], sir, body['price'], body['ami'], body['type'], body['sg'], body['subnet'])
+            update_instance_record(body['name'], sir, body['price'], body['ami'], body['type'], body['sg'],
+                                   body['subnet'], ns_record=body['ns_record'])
             logger.info('deleting message.')
             delete_message(receipt_handle)
         else:
@@ -184,35 +202,40 @@ def handle_incomming_message(body, receipt_handle, timestamp):
 
 
 # request_new_spot('0.05', 'ami-0fc8c8e37cd7db658', 'm5a.large', 'sg-056964e89bbf05266', 'subnet-03fb65f37827a8971'):
-def request_new_spot(spot_price, ami, type, sg, subnet):
-    userdata_script = f"""#!/bin/bash
-      # DONT EXECUTE !
-      #
-      # This script will be executed only once on instance first launch
-      # do here:
-      #
-      # 1. local config changes
-      # 2. local services startup (may not be necessary)
-      # 3. change DNS A records using route53 (mat be done externally also), if local need aws credentials
-      # 4. change prompt (lms/root)
-      #
-      LOG=/home/lms/user-data.output
-      echo "$(date) - starting"           >> $LOG
-      echo "executing userdata.sh"        >> $LOG
-      echo "sourcing credentials"         >> $LOG
-      source /home/lms/sqs-config
-      echo "publish metadata to SQS       >> $LOG
-      python3 /home/lms/publish-metadata.py         >> $LOG
-      echo "starting self redeployment"   >> $LOG
-      /home/lms/self-deploy.sh            >> $LOG
-      echo "$(date) - done."              >> $LOG
-      echo "waiting 2m"                   >> $LOG
-      sleep 2m
-      echo "waiting 10s before reboot"    >> $LOG
-      sleep 10s
-      reboot
-    """
-    userdata_script = base64.b64encode(userdata_script.encode("ascii")).decode("ascii")
+def request_new_spot(name, spot_price, ami, _type, sg, subnet):
+    userdata_script = """#!/bin/bash
+
+#
+# DONT EXECUTE !
+#
+# This script will be executed only once on instance first launch
+# do here:
+#
+# 1. local config changes
+# 2. local services startup (may not be necessary)
+# 3. change DNS A records using route53 (mat be done externally also), if local need aws credentials
+# 4. change prompt (lms/root)
+
+LOG=/home/lms/user-data.output
+echo "$(date) - starting"              >> $LOG
+echo "executing userdata.sh"           >> $LOG
+echo "sourcing credentials"            >> $LOG
+source /home/lms/sqs-config
+echo "publish metadata to SQS"         >> $LOG
+python3 /home/lms/publish-metadata.py  >> $LOG
+echo "starting self redeployment"      >> $LOG
+/home/lms/self-deploy.sh               >> $LOG
+echo "$(date) - done."                 >> $LOG
+#echo "waiting 2m"                     >> $LOG
+#sleep 2m
+echo "waiting 20s before reboot"       >> $LOG
+sleep 20s
+reboot
+"""
+    userdata_base64 = base64.b64encode(userdata_script.encode("ascii")).decode("ascii")
+
+    log('requesting a new spot instance for {} [price: {}. ami: {}, type: {}]'
+        .format(name, spot_price, ami, _type))
 
     response = client.request_spot_instances(
         DryRun=False,
@@ -223,8 +246,8 @@ def request_new_spot(spot_price, ami, type, sg, subnet):
         Type='one-time',
         LaunchSpecification={
             'ImageId': ami,
-            'InstanceType': type,
-            'UserData': userdata_script,
+            'InstanceType': _type,
+            'UserData': userdata_base64,
             'NetworkInterfaces': [
                 {
                     'AssociatePublicIpAddress': True,
@@ -235,7 +258,36 @@ def request_new_spot(spot_price, ami, type, sg, subnet):
             ]
         }
     )
-    x=1
+    return response
+
+
+def change_ns_record(name, ns_record, public_ip):
+
+    hosted_zone_id = '/hostedzone/ZSPS4JHX4X90T'
+    # ns_record = 'lms-chen.lms.lumosglobal.com'
+    # public_ip = '7.7.7.7'
+
+    response = route53_client.change_resource_record_sets(
+        HostedZoneId=hosted_zone_id,
+        ChangeBatch={
+            'Comment': 'string',
+            'Changes': [
+                {
+                    'Action': 'UPSERT',
+                    'ResourceRecordSet': {
+                        'Name': ns_record,
+                        'Type': 'A',
+                        'TTL': 123,
+                        'ResourceRecords': [
+                            {
+                                'Value': public_ip
+                            },
+                        ]
+                    }
+                }
+            ]
+        }
+    )
     return response
